refactor(canvas): #735 CardFrame god card を分割し selector pure 違反を修正
`AgentNodeCard/CardFrame.tsx` (~900 行の god card) で presentation /
handoff / inject / summary が一体化し、さらに zustand selector callback
内で ref を mutate していた pure 違反を解消する。

selector pure 違反の修正:
- 旧実装は `useCanvasStore((s) => { ...; lastTeamMembersSigRef.current =
  sig; return sig; })` と selector callback 内で ref を mutate していた。
  selector は pure 前提 (zustand が equality bailout / StrictMode 二重実行)
  なので前提違反だった。
- signature 計算を `useSyncExternalStore` ベースの新 hook
  `useTeamMembersSig` へ移し、selector からの ref mutate を撤廃。
  getSnapshot は「未変更ならキャッシュ値を返す」React 公式の規定動作で、
  drag 中の O(N) ループスキップもこの枠内で表現する。

god card の分割 (CardFrame.tsx は ~900 → 429 行):
- `CardPresentation.tsx` — ヘッダーの視覚表現 (avatar / title / role
  badge / status pill / close ボタン)。
- `CardHandoff.tsx` — Leader 専用 handoff 作成ボタン + bracketed-paste
  注入フロー (handoffBusy state とロジックを内包)。
- `CardInject.tsx` — PTY inject 失敗の警告 row + リトライ UI
  (retryBusy state + team:inject_failed 購読を内包)。
- `CardSummary.tsx` — current task / 経過 / health / 未読 inbox サマリ
  (純粋な表示コンポーネント、値は親が算出して props で渡す)。
- `CardFrame.tsx` は root として state/派生の解決 + layout (NodeResizer /
  Handle / .canvas-agent-card) + 子コンポーネントの合成に専念。

DOM 構造・クラス名・イベント・トーストはすべて元 CardFrame と一致。
振る舞いは不変の純粋なリファクタ。

Closes #735
