+++ src/renderer/src/components/canvas/cards/AgentNodeCard/CardFrame.tsx
@@ -1,16 +1,25 @@
 /**
- * AgentNodeCard / CardFrame
+ * AgentNodeCard / CardFrame (root)
  *
  * Issue #487: AgentNodeCard 単一ファイルから「カード枠 + role visual + handoff UI」
  * を切り出したもの。pty / xterm の配線は隣接の TerminalOverlay.tsx に分けてある。
  *
- * 責務:
+ * Issue #735: ~900 行の god card だった本ファイルを以下の子コンポーネントへ分割した。
+ *   - `CardFrame.tsx`        — root: state/派生の解決 + layout (NodeResizer / Handle /
+ *                              .canvas-agent-card) + 子コンポーネントの合成
+ *   - `CardPresentation.tsx` — ヘッダーの視覚表現 (avatar / title / role / status pill)
+ *   - `CardHandoff.tsx`      — Leader 専用 handoff 作成ボタン + 注入フロー
+ *   - `CardInject.tsx`       — PTY inject 失敗の警告 row + リトライ UI
+ *   - `CardSummary.tsx`      — current task / 経過 / health / 未読 inbox サマリ
+ * さらに、zustand selector callback 内で ref を mutate していた pure 違反を
+ * `useTeamMembersSig` (useSyncExternalStore ベース) へ移して解消した。
+ *
+ * 責務 (この root):
  *   - NodeResizer + 入出力 Handle (xyflow)
  *   - ロール由来の accent / avatar / 表示ラベル (resolveAgentVisual)
- *   - ヘッダー (title / organization / role / StatusBadge / handoff button / close)
- *   - Leader 専用の handoff 作成 → bracketed paste 注入フロー
  *   - 起動引数 (command / args / sysPrompt / codexInstructions) の解決
- *   - TerminalOverlay の mount & 共有 ref / callback の橋渡し
+ *   - agent-activity store への activity / summary の publish
+ *   - 子コンポーネント (Presentation / Handoff / Inject / Summary) と TerminalOverlay の合成
  *
  * 挙動は元 AgentNodeCard.tsx と完全一致。構造のみ整理。
  */
@@ -22,30 +31,17 @@ import {
   type Node,
   type NodeProps
 } from '@xyflow/react';
-import {
-  AlertTriangle,
-  ClipboardCheck,
-  ClipboardList,
-  Clock,
-  Heart,
-  HeartPulse,
-  Inbox,
-  RotateCcw,
-  Skull
-} from 'lucide-react';
 import { useT } from '../../../../lib/i18n';
 import { useTeamHealth } from '../../../../lib/use-team-health';
-import { deriveHealth, type HealthState } from '../../../../lib/agent-health';
-import { useTeamInjectFailed } from '../../../../lib/use-team-inject-failed';
-import { useTeamInboxRead } from '../../../../lib/use-team-inbox-read';
+import { deriveHealth } from '../../../../lib/agent-health';
 import { useTeamHandoff } from '../../../../lib/use-team-handoff';
+import { useTeamInboxRead } from '../../../../lib/use-team-inbox-read';
 import { applyHandoffArrival, applyInboxRead } from './unread-inbox-count';
 import { useSettings } from '../../../../lib/settings-context';
 import {
   useCanvasStore,
   NODE_MIN_W,
   NODE_MIN_H,
-  agentPayloadOf,
   type CardDataOf
 } from '../../../../stores/canvas';
 import { useAgentActivityStore } from '../../../../stores/agent-activity';
@@ -60,136 +56,16 @@ import { resolveAgentConfig } from '../../../../lib/agent-resolver';
 import { useToast } from '../../../../lib/toast-context';
 import {
   deriveCardSummary,
-  type CardSummary
+  type CardSummary as CardSummaryData
 } from '../../../../lib/agent-summary';
 import type { TerminalViewHandle } from '../../../TerminalView';
-import type {
-  HandoffCheckpoint,
-  HandoffReference
-} from '../../../../../../types/shared';
 import type { AgentPayload, AgentStatus } from './types';
 import { TerminalOverlay } from './TerminalOverlay';
-
-/**
- * pty 起動時の status 文字列 ("実行中: claude --append-system-prompt ...long text...") を
- * 最初のフラグ/引数まで切り詰める。チームプロンプトなど巨大な文字列がヘッダに溢れるのを防ぐ。
- */
-function shortStatus(s: string): string {
-  // "実行中: claude --append-system-prompt あなたは..." → "実行中: claude"
-  const m = s.match(/^(\S+:\s*)?([^\s]+)/);
-  if (m) return `${m[1] ?? ''}${m[2]}`;
-  return s.length > 32 ? s.slice(0, 32) + '…' : s;
-}
-
-/**
- * 絶対パスからファイル名だけを返す。Windows (`\`) と POSIX (`/`) の両方に対応するため
- * path モジュールを使わず手元で処理する (renderer 側に node:path は無い)。
- */
-function basenameOf(absPath: string): string {
-  const normalized = absPath.replace(/\\/g, '/');
-  const idx = normalized.lastIndexOf('/');
-  return idx >= 0 ? normalized.slice(idx + 1) : normalized;
-}
-
-function handoffReferenceOf(
-  handoff: HandoffCheckpoint | HandoffReference
-): HandoffReference {
-  return {
-    id: handoff.id,
-    kind: handoff.kind,
-    status: handoff.status,
-    createdAt: handoff.createdAt,
-    updatedAt: handoff.updatedAt,
-    jsonPath: handoff.jsonPath,
-    markdownPath: handoff.markdownPath,
-    fromAgentId: handoff.fromAgentId,
-    toAgentId: handoff.toAgentId,
-    replacementForAgentId: handoff.replacementForAgentId
-  };
-}
-
-/**
- * Issue #423: Leader 自身に「引き継ぎ手順」を伝える PTY 注入用プロンプトを組み立てる。
- * UI 側で handoff document を保存した直後、保存先パスをこのプロンプトに埋めて Leader の
- * PTY に bracketed paste で注入する。Leader は MCP `team_create_leader` → `team_switch_leader`
- * を呼び、自律的に新 Leader へ交代する。
- */
-function buildLeaderHandoffPrompt(markdownPath: string, handoffId: string): string {
-  return [
-    '【引き継ぎ手順】',
-    '',
-    `引き継ぎ書を保存しました: ${markdownPath}`,
-    `Handoff id: ${handoffId}`,
-    '',
-    '次の手順で引き継ぎを完了してください:',
-    '1. 上記 handoff markdown を Read tool で読み、現在の作業状況・未完了タスク・次アクションを確認する。',
-    '2. 必要なら handoff の Notes / Next Actions を補強する追加メモを書き足す。',
-    '3. MCP tool `team_create_leader` を呼び、新しい Leader を採用する:',
-    `     team_create_leader({ handoff_id: "${handoffId}" })`,
-    '   返り値の `agentId` を控えること。',
-    '4. 新 Leader が起動したら、`team_send` で agentId 宛にこの handoff のパスと「お前が新 Leader だ」という旨を伝える:',
-    `     team_send({ to: "<上で得た agentId>", handoff_id: "${handoffId}", message: "あなたが新 Leader です。handoff を読んで team_ack_handoff({ handoff_id: '${handoffId}' }) を呼び、ACK を返してください: ${markdownPath}" })`,
-    '5. 新 Leader が `team_ack_handoff` と ACK を返したら、MCP tool `team_switch_leader` を呼ぶ:',
-    `     team_switch_leader({ new_leader_agent_id: "<上で得た agentId>", handoff_id: "${handoffId}" })`,
-    '   呼び出し成功後、約 2 秒で自分のカードが自動的に閉じられる。',
-    '',
-    '上記を順に実行してください。'
-  ].join('\n');
-}
-
-/** 文字列を bracketed paste マーカーで包む。Claude/Codex TUI に「1 件のペースト」として渡る。 */
-function wrapBracketedPaste(text: string): string {
-  return `\x1b[200~${text}\x1b[201~`;
-}
-
-/**
- * Issue #521: deriveCardSummary が返す `{ unit, value }` を i18n キーに変換する。
- * unit が 'now' の時は値を埋め込まないキー、それ以外は `{value}` パラメータを渡す。
- * lastOutputAgo が null (= 起動直後で未観測) のときは「観測なし」のキーへフォールバック。
- */
-function formatAgoLabel(
-  ago: { unit: 'now' | 'sec' | 'min' | 'hour' | 'day'; value: number } | null,
-  t: (key: string, params?: Record<string, string | number>) => string
-): string {
-  if (ago === null) return t('agentCard.summary.ago.unobserved');
-  if (ago.unit === 'now') return t('agentCard.summary.ago.now');
-  return t(`agentCard.summary.ago.${ago.unit}`, { value: ago.value });
-}
-
-/**
- * Issue #510: 「●alive | ◐stale | ○dead」 + 経過秒/分 + 自己申告ステータスを 1 行に整形する。
- * - alive: status があれば status を出す。なければ 'alive' のみ。
- * - stale / dead: 経過時間を強調 ('沈黙 N 分')。
- */
-function formatHealthLabel(
-  state: HealthState,
-  ageMs: number | null,
-  currentStatus: string | null,
-  t: (key: string, params?: Record<string, string | number>) => string
-): string {
-  const stateLabel = t(`agentCard.summary.health.state.${state}`);
-  if (state === 'alive') {
-    if (currentStatus && currentStatus.trim().length > 0) {
-      const status = currentStatus.length > 32 ? currentStatus.slice(0, 31) + '…' : currentStatus;
-      return `${stateLabel} · ${status}`;
-    }
-    return stateLabel;
-  }
-  if (ageMs === null) return stateLabel;
-  // 沈黙時間: 1 分未満は秒、それ以上は分単位 (停滞は「N 分」が直感的)。
-  const sec = Math.floor(ageMs / 1000);
-  if (sec < 60) {
-    return t('agentCard.summary.health.silent.sec', {
-      state: stateLabel,
-      value: sec
-    });
-  }
-  const min = Math.floor(sec / 60);
-  return t('agentCard.summary.health.silent.min', {
-    state: stateLabel,
-    value: min
-  });
-}
+import { useTeamMembersSig } from './use-team-members-sig';
+import { CardPresentation } from './CardPresentation';
+import { CardHandoff } from './CardHandoff';
+import { CardInject } from './CardInject';
+import { CardSummary, type CardSummaryHealth } from './CardSummary';
 
 // Issue #732: `NodeProps` を `Node<CardDataOf<'agent'>>` で具体化することで
 // `data.payload` が `AgentPayload` として読め、`unknown` からの inline cast が不要になる。
@@ -215,7 +91,6 @@ function AgentNodeCardImpl({
   const accent = visual.agentAccent;
   const organizationAccent = visual.organizationAccent;
   const title = data?.title ?? visual.label;
-  const [handoffBusy, setHandoffBusy] = useState(false);
   const [status, setStatus] = useState<string>('');
   const [activity, setActivityState] = useState<AgentStatus>('idle');
 
@@ -253,25 +128,13 @@ function AgentNodeCardImpl({
   // パフォーマンス: 旧実装は `useCanvasStore((s) => s.nodes)` で全 nodes を購読していたため、
   // ノードを 1 ピクセル動かすだけで全 AgentNodeCard が再レンダーし Canvas が重かった。
   // 対策として primitive な signature 文字列 (agentId|role|agent を ; で連結) を購読し、
-  // 文字列 equality で React がデフォルトで bailout できるようにする。drag/resize では
-  // signature が変わらないので再レンダーが発生しない。
-  const lastTeamMembersSigRef = useRef('');
-  const teamMembersSig = useCanvasStore((s) => {
-    if (!payload.teamId) return '';
-    if (s.isDragging) return lastTeamMembersSigRef.current;
-    const sigs: string[] = [];
-    for (const n of s.nodes) {
-      if (n.type !== 'agent') continue;
-      // Issue #732: 旧 `n.data?.payload as AgentPayload` を agentPayloadOf に置換。
-      const p = agentPayloadOf(n.data);
-      const rp = p?.roleProfileId ?? p?.role;
-      if (!p || p.teamId !== payload.teamId || !p.agentId || !rp) continue;
-      sigs.push(`${p.agentId}:${rp}:${p.agent ?? 'claude'}`);
-    }
-    const sig = sigs.join(';');
-    lastTeamMembersSigRef.current = sig;
-    return sig;
-  });
+  // 文字列 equality で React がデフォルトで bailout できるようにする。
+  //
+  // Issue #735: 旧実装は signature 計算を zustand selector callback 内で行いつつ
+  // `lastTeamMembersSigRef.current = sig` と ref を mutate していた (selector pure 違反)。
+  // signature 計算は useSyncExternalStore ベースの `useTeamMembersSig` へ移し、
+  // selector からの ref mutate を撤廃した。
+  const teamMembersSig = useTeamMembersSig(payload.teamId);
   const teamMembers = useMemo(() => {
     if (!payload.teamId) return null;
     if (teamMembersSig === '')
@@ -388,202 +251,6 @@ function AgentNodeCardImpl({
 
   const codexInstructions = isCodex ? sysPrompt : undefined;
 
-  // Issue #375 / #423: createHandoff は副作用 (handoff の保存 + payload.latestHandoff 更新) のみ
-  // 行い、success toast は呼び出し側 (handleCreateHandoffClick) に任せる。
-  // 呼び出し側で「保存先のパス」を PTY 注入用に取り出すため、戻り値の HandoffCheckpoint は必須。
-  const createHandoff = useCallback(async (): Promise<HandoffCheckpoint | null> => {
-    const projectRoot = cwd || payload.cwd || '';
-    if (!projectRoot) {
-      showToast(t('handoff.error.noProject'), { tone: 'error', duration: 8000 });
-      return null;
-    }
-    const snapshot = termRef.current?.getBufferText(120) ?? '';
-    const kind = roleProfileId === 'leader' ? 'leader' : 'worker';
-    const result = await window.api.handoffs.create({
-      projectRoot,
-      teamId: payload.teamId ?? null,
-      kind,
-      fromAgentId: payload.agentId ?? null,
-      fromRole: roleProfileId,
-      fromAgent: payload.agent ?? 'claude',
-      fromTitle: title,
-      sourceSessionId: payload.resumeSessionId ?? null,
-      replacementForAgentId: payload.agentId ?? null,
-      retireAfterAck: true,
-      trigger: 'manual',
-      content: {
-        summary: `${title} (${visual.label}) の Canvas handoff。保存時点の terminal snapshot と次アクションを含みます。`,
-        decisions: ['この handoff は既存セッションを --resume せず、新しいセッションへ注入するための継続メモとして保存されました。'],
-        filesTouched: [],
-        openTasks: ['handoff markdown を読み、現在の作業目的・未完了タスク・次アクションを確認する。'],
-        risks: ['terminal snapshot は直近の表示内容ベースのため、完全な会話履歴ではありません。必要なら旧 agent / team history を確認してください。'],
-        nextActions: ['handoff を読んだら ack を返し、Next Actions に沿って作業を継続する。'],
-        verification: ['handoff 作成時点では自動検証は未実行です。'],
-        notes: [`Canvas card: ${id}`, payload.teamId ? `Team: ${payload.teamId}` : 'Standalone agent'],
-        terminalSnapshot: snapshot.slice(-16_000) || null
-      }
-    });
-    if (!result.ok || !result.handoff) {
-      throw new Error(result.error ?? 'handoff create failed');
-    }
-    setCardPayload(id, { latestHandoff: handoffReferenceOf(result.handoff) });
-    return result.handoff;
-  }, [
-    cwd,
-    id,
-    visual.label,
-    payload.agent,
-    payload.agentId,
-    payload.cwd,
-    payload.resumeSessionId,
-    payload.teamId,
-    roleProfileId,
-    setCardPayload,
-    showToast,
-    t,
-    title
-  ]);
-
-  // Issue #423: 旧 `startFreshFromHandoff` (UI が直接新カードを生やす) は廃止し、
-  // 「Leader 自身が MCP 経由で交代する」フローへ移行。ボタンは下記 1 本に統合。
-
-  // Issue #423: ボタン押下時のフロー
-  //   1. Rust 側 `handoffs.create` で handoff JSON / Markdown を確実に保存
-  //   2. 保存先パス + MCP 手順を Leader 自身の PTY に bracketed paste で注入
-  //   3. Leader (Claude/Codex) が `team_create_leader` → `team_send` → `team_switch_leader`
-  //      を順に叩き、自律的に新 Leader へ交代する
-  // Leader 以外のカードでは押せない (worker の引き継ぎは将来の別 issue で対応)。
-  const handleCreateHandoffClick = useCallback(() => {
-    if (handoffBusy) return;
-    if (roleProfileId !== 'leader') {
-      showToast(t('handoff.error.notLeader'), { tone: 'error', duration: 6000 });
-      return;
-    }
-    setHandoffBusy(true);
-    void createHandoff()
-      .then((handoff) => {
-        if (!handoff) return; // noProject 等は createHandoff 側で error toast を出している
-        const fileName = basenameOf(handoff.markdownPath);
-        const markdownPath = handoff.markdownPath;
-        // Leader の PTY に「引き継ぎ手順」プロンプトを bracketed paste で注入。
-        // sendCommand(text, submit=true) は末尾に \r を付けて送信するため、
-        // 全文が 1 つの paste として確定 → Claude/Codex が読み取って MCP を叩き始める。
-        try {
-          const prompt = buildLeaderHandoffPrompt(markdownPath, handoff.id);
-          termRef.current?.sendCommand(wrapBracketedPaste(prompt), true);
-        } catch (err) {
-          const detail = err instanceof Error ? err.message : String(err);
-          showToast(t('handoff.error.injectFailed', { detail }), {
-            tone: 'error',
-            duration: 8000
-          });
-          return;
-        }
-        showToast(t('handoff.created', { file: fileName }), {
-          tone: 'success',
-          duration: 8000,
-          action: {
-            label: t('handoff.action.reveal'),
-            onClick: () => {
-              void window.api.app.revealInFileManager(markdownPath).catch((err) => {
-                console.warn('[handoff] reveal failed:', err);
-              });
-            }
-          }
-        });
-      })
-      .catch((err) => {
-        console.warn('[handoff] create failed:', err);
-        const detail = err instanceof Error ? err.message : String(err);
-        showToast(t('handoff.error.createFailed', { detail }), {
-          tone: 'error',
-          duration: 8000
-        });
-      })
-      .finally(() => setHandoffBusy(false));
-  }, [createHandoff, handoffBusy, roleProfileId, showToast, t]);
-
-  // Issue #423: 旧 `startFreshFromHandoff` 経路で使っていた `handoff_ack:` 監視は撤去。
-  // 新フローでは `team_switch_leader` MCP tool が active leader 切替 + 旧カード retire を
-  // 行うため、renderer 側で ack を listen する必要は無い。
-
-  // ---------- Issue #511: inject 失敗の警告表示 + 手動リトライ ----------
-  //
-  // `team_send` (またはリトライ後の `team_send_retry_inject`) が PTY inject に失敗した
-  // 瞬間、Hub から `team:inject_failed` event が emit される。Canvas 側はそれを受けて
-  // 該当 agent の payload.lastInjectFailure に reason を書き込む → CardFrame が warning
-  // row を render する。retry button で `window.api.team.retryInject` を呼び、成功すれば
-  // payload.lastInjectFailure を undefined クリアして warning を消す。
-  const [retryBusy, setRetryBusy] = useState(false);
-  useTeamInjectFailed(
-    useCallback(
-      (evt) => {
-        if (!payload.agentId || evt.toAgentId !== payload.agentId) return;
-        setCardPayload(id, {
-          lastInjectFailure: {
-            messageId: evt.messageId,
-            reason: { code: evt.reasonCode, message: evt.reasonMessage },
-            failedAt: evt.failedAt,
-            fromRole: evt.fromRole
-          }
-        });
-      },
-      [id, payload.agentId, setCardPayload]
-    )
-  );
-  const handleRetryInject = useCallback(() => {
-    if (retryBusy) return;
-    const failure = payload.lastInjectFailure;
-    if (!failure || !payload.teamId || !payload.agentId) return;
-    setRetryBusy(true);
-    void window.api.team
-      .retryInject({
-        teamId: payload.teamId,
-        messageId: failure.messageId,
-        agentId: payload.agentId
-      })
-      .then((result) => {
-        if (result.ok) {
-          // 成功時は warning row を消す。Hub からは team:handoff event が来るので
-          // 配信成功は Canvas 側 ActivityFeed / HandoffEdge が拾う。
-          setCardPayload(id, { lastInjectFailure: undefined });
-          showToast(t('injectFailure.retrySuccess'), {
-            tone: 'success',
-            duration: 5000
-          });
-        } else {
-          // 再失敗。Hub が `team:inject_failed` を再 emit するので useTeamInjectFailed が
-          // 新しい reason を payload に書き込む (= warning row はそのまま、内容だけ更新)。
-          const reason = result.reasonCode ?? result.error ?? 'unknown';
-          showToast(t('injectFailure.retryFailed', { reason }), {
-            tone: 'error',
-            duration: 8000
-          });
-        }
-      })
-      .catch((err) => {
-        // unknown_team / unknown_message / invalid_recipient の構造化エラーはここに来る。
-        const detail = err instanceof Error ? err.message : String(err);
-        showToast(t('injectFailure.retryError', { detail }), {
-          tone: 'error',
-          duration: 8000
-        });
-      })
-      .finally(() => setRetryBusy(false));
-  }, [
-    retryBusy,
-    payload.lastInjectFailure,
-    payload.teamId,
-    payload.agentId,
-    id,
-    setCardPayload,
-    showToast,
-    t
-  ]);
-  const handleDismissInjectWarning = useCallback(() => {
-    setCardPayload(id, { lastInjectFailure: undefined });
-  }, [id, setCardPayload]);
-
   // ---------- Issue #509: 未読 inbox 数の event-driven 集計 ----------
   //
   // `team:handoff` (= delivered = inject 成功) を受けると、自分宛のメッセージは「配信済み
@@ -631,10 +298,10 @@ function AgentNodeCardImpl({
   const publishSummary = useAgentActivityStore((s) => s.setSummary);
   const [nowTick, setNowTick] = useState(() => Date.now());
   useEffect(() => {
-    const t = window.setInterval(() => setNowTick(Date.now()), 15_000);
-    return () => window.clearInterval(t);
+    const timer = window.setInterval(() => setNowTick(Date.now()), 15_000);
+    return () => window.clearInterval(timer);
   }, []);
-  const summary = useMemo<CardSummary>(
+  const summary = useMemo<CardSummaryData>(
     () =>
       deriveCardSummary({
         payload,
@@ -649,7 +316,6 @@ function AgentNodeCardImpl({
   useEffect(() => {
     publishSummary(id, summary);
   }, [id, summary, publishSummary]);
-  const summaryAgoLabel = formatAgoLabel(summary.lastOutputAgo, t);
 
   // Issue #510: TeamHub diagnostics を 5s poll し、自カードの per-agent 行から
   // health (alive / stale / dead) と現在 status / pendingInbox を抽出する。
@@ -663,6 +329,21 @@ function AgentNodeCardImpl({
   const unreadInboxCount = hasHealthRow
     ? health.pendingInboxCount
     : payload.unreadInboxCount ?? 0;
+  // CardSummary に渡す health 派生値。health 行を出すのは teamId / agentId が両方あり
+  // state が 'unknown' でないとき (= standalone カードでは出さない)。
+  const showHealthRow = Boolean(payload.agentId) && Boolean(payload.teamId) && health.state !== 'unknown';
+  const summaryHealth = useMemo<CardSummaryHealth>(
+    () => ({
+      state: health.state,
+      ageMs: health.ageMs,
+      currentStatus: health.currentStatus,
+      stalledInbound: health.stalledInbound,
+      oldestPendingInboxAgeMs: health.oldestPendingInboxAgeMs
+    }),
+    [health]
+  );
+
+  const handleClose = useCallback(() => confirmRemoveCard(id), [confirmRemoveCard, id]);
 
   return (
     <>
@@ -679,197 +360,48 @@ function AgentNodeCardImpl({
         style={{ background: accent, width: 10, height: 10 }}
       />
       <div className="canvas-agent-card" style={cardStyle}>
-        <header className="canvas-agent-card__header">
-          <span className="canvas-agent-card__title-row">
-            <span aria-hidden="true" className="canvas-agent-card__avatar">
-              {profile.visual.glyph}
-            </span>
-            <span className="canvas-agent-card__title">{title}</span>
-            {payload.organization && (
-              <span className="canvas-agent-card__organization">
-                {payload.organization.name}
-              </span>
-            )}
-            <span className="canvas-agent-card__role">{visual.label}</span>
-          </span>
-          <span className="canvas-agent-card__actions">
-            <StatusBadge state={activity} label={t(`agentStatus.${activity}`)} />
-            {status && (
-              <span className="canvas-agent-card__status" title={status}>
-                {shortStatus(status)}
-              </span>
-            )}
-            {roleProfileId === 'leader' && (
-              <button
-                type="button"
-                className="nodrag canvas-agent-card__tool"
-                onClick={handleCreateHandoffClick}
-                disabled={handoffBusy}
-                title={t('handoff.createTooltip')}
-                aria-label={t('handoff.create')}
-              >
-                <ClipboardCheck size={13} strokeWidth={1.9} />
-              </button>
-            )}
-            <button
-              type="button"
-              className="nodrag canvas-agent-card__close"
-              onClick={() => confirmRemoveCard(id)}
-              title={t('agentCard.close')}
-              aria-label={t('agentCard.close')}
-            >
-              ×
-            </button>
-          </span>
-        </header>
-        <div
-          className={
-            'canvas-agent-card__summary' +
-            (summary.needsLeaderInput
-              ? ' canvas-agent-card__summary--alert'
-              : '')
-          }
-          aria-label={t('agentCard.summary.region')}
-        >
-          <div
-            className="canvas-agent-card__summary-row canvas-agent-card__summary-row--task"
-            title={summary.taskTitle || t('agentCard.summary.noTask')}
-          >
-            <ClipboardList size={11} strokeWidth={2} aria-hidden="true" />
-            <span className="canvas-agent-card__summary-text">
-              {summary.taskTitle || t('agentCard.summary.noTask')}
-            </span>
-          </div>
-          <div className="canvas-agent-card__summary-row canvas-agent-card__summary-row--clock">
-            <Clock size={11} strokeWidth={2} aria-hidden="true" />
-            <span className="canvas-agent-card__summary-text">
-              {summaryAgoLabel}
-            </span>
-          </div>
-          {summary.needsLeaderInput ? (
-            <div
-              className="canvas-agent-card__summary-row canvas-agent-card__summary-row--leader"
-              role="status"
-            >
-              <AlertTriangle size={11} strokeWidth={2} aria-hidden="true" />
-              <span className="canvas-agent-card__summary-text">
-                {t('agentCard.summary.needsLeader')}
-              </span>
-            </div>
-          ) : null}
-          {/* Issue #510: 自カードに対応する TeamHub diagnostics 行から health badge を表示する。 */}
-          {/* teamId / agentId が無いスタンドアロンカードでは何も出さない (= state==='unknown' は描画しない)。 */}
-          {payload.agentId && payload.teamId && health.state !== 'unknown' ? (
-            <div
-              className={
-                'canvas-agent-card__summary-row canvas-agent-card__summary-row--health' +
-                ' canvas-agent-card__summary-row--health-' + health.state
-              }
-              role="status"
-              title={
-                t('agentCard.summary.health.tooltip', {
-                  state: t(`agentCard.summary.health.state.${health.state}`),
-                  status: health.currentStatus ?? t('agentCard.summary.health.noStatus')
-                })
-              }
-            >
-              {health.state === 'alive' ? (
-                <Heart size={11} strokeWidth={2.2} aria-hidden="true" />
-              ) : health.state === 'stale' ? (
-                <HeartPulse size={11} strokeWidth={2.2} aria-hidden="true" />
-              ) : (
-                <Skull size={11} strokeWidth={2.2} aria-hidden="true" />
-              )}
-              <span className="canvas-agent-card__summary-text">
-                {formatHealthLabel(health.state, health.ageMs, health.currentStatus, t)}
-              </span>
-            </div>
-          ) : null}
-          {/* Issue #509: 配送済みだが team_read で確認していない message の数。 */}
-          {/* 60s 超過で stalled クラスを追加して警告色に切り替える。 */}
-          {unreadInboxCount > 0 ? (() => {
-            const ageMs = hasHealthRow
-              ? health.oldestPendingInboxAgeMs ?? 0
-              : payload.oldestUnreadDeliveredAt
-                ? Math.max(0, nowTick - new Date(payload.oldestUnreadDeliveredAt).getTime())
-                : 0;
-            const stalled = hasHealthRow ? health.stalledInbound : ageMs >= 60_000;
-            const ageSec = Math.floor(ageMs / 1000);
-            return (
-              <div
-                className={
-                  'canvas-agent-card__summary-row canvas-agent-card__summary-row--unread' +
-                  (stalled
-                    ? ' canvas-agent-card__summary-row--unread-stalled'
-                    : '')
-                }
-                role="status"
-                title={t('inboxUnread.tooltip', {
-                  count: unreadInboxCount,
-                  ageSec
-                })}
-              >
-                <Inbox size={11} strokeWidth={2} aria-hidden="true" />
-                <span className="canvas-agent-card__summary-text">
-                  {t('inboxUnread.label', {
-                    count: unreadInboxCount,
-                    ageSec
-                  })}
-                </span>
-              </div>
-            );
-          })() : null}
-        </div>
-        {/* Issue #511: PTY inject 失敗 warning row。 */}
-        {/* 通常時は何も rendering されず、`team:inject_failed` が来た瞬間に出現する。 */}
-        {/* `__summary` block の sibling として置き、既存 header の flex レイアウトを破壊しない。 */}
-        {payload.lastInjectFailure ? (
-          <div
-            className="canvas-agent-card__inject-warning"
-            role="alert"
-            aria-live="polite"
-          >
-            <AlertTriangle
-              size={12}
-              strokeWidth={2}
-              className="canvas-agent-card__inject-warning__icon"
-              aria-hidden="true"
+        <CardPresentation
+          cardId={id}
+          title={title}
+          roleLabel={visual.label}
+          glyph={profile.visual.glyph}
+          organizationName={payload.organization?.name}
+          activity={activity}
+          status={status}
+          handoff={
+            <CardHandoff
+              cardId={id}
+              payload={payload}
+              roleProfileId={roleProfileId}
+              title={title}
+              visualLabel={visual.label}
+              cwd={cwd}
+              termRef={termRef}
+              setCardPayload={setCardPayload}
+              showToast={showToast}
+              t={t}
             />
-            <span
-              className="canvas-agent-card__inject-warning__text"
-              title={payload.lastInjectFailure.reason.message}
-            >
-              {t('injectFailure.title', {
-                code: payload.lastInjectFailure.reason.code,
-                message: payload.lastInjectFailure.reason.message
-              })}
-            </span>
-            <button
-              type="button"
-              className="nodrag canvas-agent-card__inject-warning__retry"
-              onClick={handleRetryInject}
-              disabled={retryBusy}
-              title={t('injectFailure.retry')}
-              aria-label={t('injectFailure.retry')}
-            >
-              <RotateCcw size={11} strokeWidth={2} aria-hidden="true" />
-              <span>
-                {retryBusy
-                  ? t('injectFailure.retryBusy')
-                  : t('injectFailure.retry')}
-              </span>
-            </button>
-            <button
-              type="button"
-              className="nodrag canvas-agent-card__inject-warning__dismiss"
-              onClick={handleDismissInjectWarning}
-              title={t('injectFailure.dismiss')}
-              aria-label={t('injectFailure.dismiss')}
-            >
-              ×
-            </button>
-          </div>
-        ) : null}
+          }
+          onClose={handleClose}
+          t={t}
+        />
+        <CardSummary
+          summary={summary}
+          health={summaryHealth}
+          showHealthRow={showHealthRow}
+          hasHealthRow={hasHealthRow}
+          unreadInboxCount={unreadInboxCount}
+          oldestUnreadDeliveredAt={payload.oldestUnreadDeliveredAt}
+          nowTick={nowTick}
+          t={t}
+        />
+        <CardInject
+          cardId={id}
+          payload={payload}
+          setCardPayload={setCardPayload}
+          showToast={showToast}
+          t={t}
+        />
         <TerminalOverlay
           cardId={id}
           termRef={termRef}
@@ -891,26 +423,6 @@ function AgentNodeCardImpl({
         style={{ background: accent, width: 10, height: 10 }}
       />
     </>
-  );
-}
-
-/** ヘッダー右の小さなステータスドット (idle=灰, thinking=黄, typing=accent パルス) */
-function StatusBadge({
-  state,
-  label
-}: {
-  state: AgentStatus;
-  label: string;
-}): JSX.Element {
-  return (
-    <span
-      title={label}
-      aria-label={label}
-      className={`canvas-agent-status canvas-agent-status--${state}`}
-    >
-      <span className="canvas-agent-status__dot" />
-      <span>{label}</span>
-    </span>
   );
 }
